rest-api(api): fold templated-OS instance tests into instance_test.go
